docs: expand README with usage
