+++ nethira/analysis/__init__.py
@@ -1,0 +1,1 @@
+"Analysis modules for Nethira."
