+++ nethira/models/device_info.py
