+++ nethira/utils/__init__.py
@@ -1,0 +1,5 @@
+"Utility helpers for Nethira."
+
+from .adb_utils import get_adb_path
+
+__all__ = ["get_adb_path"]
