+++ tools/clean_up_project.bat
